+++ .gitignore
@@ -1,0 +1,5 @@
+# Dependency directory
+node_modules
+
+# Compiled JS directory
+lib
