+++ dev/index.html
@@ -2,7 +2,7 @@
 <html lang="en" data-framework="typescript">
   <head>
     <meta charset="utf-8">
-    <title>OPDS Viewer Prototype</title>
+    <title>OPDS Browser</title>
     <script src="http://localhost:8090/webpack-dev-server.js"></script>
     <style>
       body { margin: 0; }
