+++ tsd.json
@@ -22,12 +22,6 @@
     },
     "react/react-addons-test-utils.d.ts": {
       "commit": "86dbea8fc37d9473fee465da4f0a21bea4f8cbd9"
-    },
-    "form-data/form-data.d.ts": {
-      "commit": "86dbea8fc37d9473fee465da4f0a21bea4f8cbd9"
-    },
-    "request/request.d.ts": {
-      "commit": "86dbea8fc37d9473fee465da4f0a21bea4f8cbd9"
     }
   }
 }
