+++ typings/tsd.d.ts
@@ -1,8 +1,6 @@
-/// <reference path="react/react.d.ts" />
-/// <reference path="react/react-dom.d.ts" />
-/// <reference path="lodash/lodash.d.ts" />
 /// <reference path="jest/jest.d.ts" />
+/// <reference path="lodash/lodash.d.ts" />
 /// <reference path="node/node.d.ts" />
 /// <reference path="react/react-addons-test-utils.d.ts" />
-/// <reference path="form-data/form-data.d.ts" />
-/// <reference path="request/request.d.ts" />
+/// <reference path="react/react-dom.d.ts" />
+/// <reference path="react/react.d.ts" />
